Move runner to own package
